Extract classes to files and move things around a bit
