+++ posts/src/main/java/com/neaniesoft/vermilion/posts/data/PostToLinkConverter.kt
@@ -1,24 +1,14 @@
 package com.neaniesoft.vermilion.posts.data
 
 import android.net.Uri
-import com.github.michaelbull.result.Result
-import com.github.michaelbull.result.map
-import com.github.michaelbull.result.mapError
-import com.github.michaelbull.result.runCatching
 import com.neaniesoft.vermilion.api.entities.Awarding
 import com.neaniesoft.vermilion.api.entities.Link
-import com.neaniesoft.vermilion.api.entities.LinkThing
-import com.neaniesoft.vermilion.posts.data.http.PostsService
-import com.neaniesoft.vermilion.posts.domain.entities.AfterKey
 import com.neaniesoft.vermilion.posts.domain.entities.AuthorName
 import com.neaniesoft.vermilion.posts.domain.entities.Award
 import com.neaniesoft.vermilion.posts.domain.entities.AwardCount
 import com.neaniesoft.vermilion.posts.domain.entities.AwardName
-import com.neaniesoft.vermilion.posts.domain.entities.BeforeKey
 import com.neaniesoft.vermilion.posts.domain.entities.CommentCount
-import com.neaniesoft.vermilion.posts.domain.entities.Community
 import com.neaniesoft.vermilion.posts.domain.entities.CommunityName
-import com.neaniesoft.vermilion.posts.domain.entities.FrontPage
 import com.neaniesoft.vermilion.posts.domain.entities.ImagePostSummary
 import com.neaniesoft.vermilion.posts.domain.entities.LinkHost
 import com.neaniesoft.vermilion.posts.domain.entities.NamedCommunity
@@ -28,68 +18,16 @@ import com.neaniesoft.vermilion.posts.domain.entities.PostId
 import com.neaniesoft.vermilion.posts.domain.entities.PostSummary
 import com.neaniesoft.vermilion.posts.domain.entities.PostTitle
 import com.neaniesoft.vermilion.posts.domain.entities.PreviewText
-import com.neaniesoft.vermilion.posts.domain.entities.ResultSet
 import com.neaniesoft.vermilion.posts.domain.entities.Score
 import com.neaniesoft.vermilion.posts.domain.entities.TextPostSummary
 import com.neaniesoft.vermilion.posts.domain.entities.UriImage
 import com.neaniesoft.vermilion.posts.domain.entities.VideoPostSummary
-import com.neaniesoft.vermilion.posts.domain.errors.PostError
-import com.neaniesoft.vermilion.posts.domain.errors.PostsApiError
-import com.neaniesoft.vermilion.utils.logger
+import com.neaniesoft.vermilion.utils.anonymousLogger
 import org.apache.commons.text.StringEscapeUtils
 import java.net.URL
 import java.time.Instant
 import java.util.Locale
-import javax.inject.Inject
-import javax.inject.Singleton
 import kotlin.math.roundToLong
-
-@Singleton
-class PostRepositoryImpl @Inject constructor(
-    private val postsService: PostsService
-) : PostRepository {
-
-    private val logger by logger()
-
-    override suspend fun postsForCommunity(
-        community: Community,
-        requestedCount: Int,
-        previousCount: Int?,
-        afterKey: String?
-    ): Result<ResultSet<Post>, PostError> {
-        logger.debugIfEnabled { "Loading posts for $community" }
-        return runCatching {
-            when (community) {
-                is FrontPage -> {
-                    postsService.frontPageBest(
-                        requestedCount,
-                        null,
-                        afterKey,
-                        previousCount
-                    )
-                }
-                is NamedCommunity -> TODO()
-            }
-        }.mapError {
-            PostsApiError(it)
-        }.map { response ->
-            val posts = response.data.children.mapNotNull { child ->
-                if (child is LinkThing) {
-                    child.data.toPost()
-                } else {
-                    logger.warnIfEnabled { "Unknown thing type in posts response" }
-                    null
-                }
-            }
-
-            ResultSet(
-                posts,
-                response.data.before?.let { BeforeKey(it) },
-                response.data.after?.let { AfterKey(it) },
-            )
-        }
-    }
-}
 
 internal fun Link.toPost(): Post {
     return Post(
@@ -144,6 +82,8 @@ internal fun Link.postSummary(): PostSummary {
             )
         }
         else -> {
+            val logger by anonymousLogger("postSummary()")
+            logger.debugIfEnabled { "Unrecognised post hint ($hint), defaulting to text post type" }
             TextPostSummary(PreviewText(""))
         }
     }
