--- posts/src/main/java/com/neaniesoft/vermilion/posts/data/entities/PagingQuery.kt
@@ -1,8 +1,0 @@
-package com.neaniesoft.vermilion.posts.data.entities
-
-import com.neaniesoft.vermilion.posts.domain.entities.ListingKey
-
-data class PagingQuery(
-    val listingKey: ListingKey,
-    val previousCount: Int
-)
