+++ posts/src/main/java/com/neaniesoft/vermilion/posts/domain/entities/Post.kt
@@ -16,6 +16,3 @@ data class Post(
     val flags: Set<PostFlags>,
     val link: URL
 )
-
-@JvmInline
-value class PostId(val value: String)
