+++ posts/src/main/java/com/neaniesoft/vermilion/posts/domain/entities/PostId.kt
@@ -1,0 +1,4 @@
+package com.neaniesoft.vermilion.posts.domain.entities
+
+@JvmInline
+value class PostId(val value: String)
