+++ posts/src/main/java/com/neaniesoft/vermilion/posts/ui/PostCard.kt
@@ -1,27 +1,19 @@
 package com.neaniesoft.vermilion.posts.ui
 
-import android.net.Uri
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.BoxWithConstraints
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.material.Card
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import coil.compose.rememberImagePainter
 import com.neaniesoft.vermilion.posts.R
 import com.neaniesoft.vermilion.posts.domain.entities.AuthorName
 import com.neaniesoft.vermilion.posts.domain.entities.CommentCount
@@ -107,56 +99,11 @@ fun PostCardPlaceholder() {
     }
 }
 
-@Composable
-fun TextSummary(content: String) {
-    Text(
-        text = content,
-        style = MaterialTheme.typography.body1
-    )
-}
-
-@Composable
-fun ImageSummary(imageUri: Uri) {
-    val painter = rememberImagePainter(imageUri.toString())
-
-    BoxWithConstraints(Modifier.fillMaxWidth()) {
-        Image(modifier = Modifier.size(maxWidth), painter = painter, contentDescription = "")
-    }
-}
-
-@Composable
-fun VideoSummary(previewImageUri: Uri) {
-    val painter = rememberImagePainter(previewImageUri.toString())
-
-    BoxWithConstraints(Modifier.fillMaxWidth()) {
-        Image(
-            modifier = Modifier.size(maxWidth),
-            painter = painter,
-            contentDescription = ""
-        )
-        Image(
-            modifier = Modifier
-                .size(48.dp)
-                .align(Alignment.Center),
-            imageVector = ImageVector.vectorResource(id = R.drawable.ic_baseline_play_circle_filled_24),
-            contentDescription = "Video icon"
-        )
-    }
-}
-
 @Preview(name = "Text Post card light")
 @Composable
 fun PostCardPreview() {
     VermilionTheme {
         PostCard(post = DUMMY_TEXT_POST)
-    }
-}
-
-@Preview(name = "Video summary")
-@Composable
-fun VideoSummaryPreview() {
-    VermilionTheme {
-        VideoSummary(previewImageUri = Uri.parse("https://www.google.com.au/images/branding/googlelogo/2x/googlelogo_color_272x92dp.png"))
     }
 }
