+++ utils/src/main/java/com/neaniesoft/vermilion/utils/AndroidLogger.kt
@@ -38,4 +38,12 @@ class LoggerDelegate(private val level: Logger.Level) {
     }
 }
 
+class AnonymousLoggerDelegate(private val level: Logger.Level, private val tag: String) {
+    operator fun getValue(thisRef: Nothing?, property: KProperty<*>): Logger {
+        return AndroidLogger(tag, level)
+    }
+}
+
 fun logger(level: Logger.Level = Logger.Level.DEBUG) = LoggerDelegate(level)
+fun anonymousLogger(tag: String, level: Logger.Level = Logger.Level.DEBUG) =
+    AnonymousLoggerDelegate(level, tag)
